background/requirements-bg.txt
aiodns==3.2.0
aioping==0.4.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncping3==3.1.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
idna==3.10
numpy==2.1.2
opencv-python==4.10.0.84
pycares==4.4.0
pycparser==2.22
pydot==3.0.2
pyparsing==3.2.0
pytz==2024.2
rel==0.4.9.19
requests==2.32.3
six==1.16.0
tzlocal==5.2
urllib3==2.2.3
websocket-client==1.8.0
